Call repo names as names instead of urls.
I plan to use github urls, but I like the word names better since
nothing about how I'm using the string makes special use of it as a
github url. As far as syncer is concerned, a repo name is nothing more
than an identifying string it can use to know which files are at home in
which paths.