Add transitive closure to diff checking.
This is a feature I've looked forward to for a while as a user. It means that
any file changes enacted by a check command also give the user an explicit
opportunity to be propagated forward throughout their graph of file connections.
In other words, unless the user choose to stop a propagation, a changed file in
any location in the graph will reach all copies of that file in a single run of
`syncer check`.

Yes, intuitively, that's what you'd want to happen, and it may seem surprising
that this wasn't already the default behavior. It was just a bit more work to
make syncer smart enough to do this for us.