v1 of full file add/delete handling.
This adds a global called _copy_dirs and a ~/.syncer/copy_dirs config file that
keep track of which directories are copies of which home directories. This also
gives the user some options to either untrack such a directory mapping or to
exclude individual files from such a mapping.

I'm concerned that this code may be buggy. It deserves more testing. I have
tested it to a small extent so far, though.