Add missing directories if appropriate.
A directory may be missing if, for example, a home repo adds subdir Y within dir
X that's tracked in copy_X. Then any files in Y will be copied into copy_X/Y,
and the copy of Y will be created if needed before such a file copy. This has
always been the intended behavior, but it implicitly assumed that all needed
directories existed before this commit.